Move resources.ts to utils/resources.ts
