+++ src/k8s-engine/images/mobyImageProcessor.ts
@@ -2,7 +2,7 @@ import { spawn } from 'child_process';
 import path from 'path';
 
 import Logging from '@/utils/logging';
-import resources from '@/resources';
+import resources from '@/utils/resources';
 import * as imageProcessor from '@/k8s-engine/images/imageProcessor';
 import mainEvents from '@/main/mainEvents';
 import * as K8s from '@/k8s-engine/k8s';
