+++ src/k8s-engine/kubectl.ts
@@ -1,9 +1,8 @@
 'use strict';
 
 import childProcess, { spawn } from 'child_process';
-import os from 'os';
 import process from 'process';
-import resources from '@/resources';
+import resources from '@/utils/resources';
 
 // The K8s JS library will get the current context but does not have the ability
 // to save the context. The current version of the package targets k8s 1.18 and
