+++ src/main/tray.ts
@@ -12,7 +12,6 @@ import { KubeConfig } from '@kubernetes/client-node';
 import * as kubectl from '@/k8s-engine/kubectl';
 import kubeconfig from '@/config/kubeconfig.js';
 import { State } from '@/k8s-engine/k8s';
-import resources from '@/resources';
 import { openPreferences } from '@/window';
 import { openDashboard } from '@/window/dashboard';
 import mainEvents from '@/main/mainEvents';
